requirements.txt
Django==2.1.4
djangorestframework==3.9.0
mysqlclient==1.3.14
pytz==2018.7
uWSGI==2.0.17.1
django-service-objects
